// libs/features/collection/billing/src/lib/shared/models/mof-contribution-request-list.ts
/**
 * Copyright GOSI. All Rights Reserved.
 * This software is the proprietary information of GOSI.
 * Use is subject to license terms.
 */
import { Page } from './page';

export class MofContributionRequestList {
  page: Page = new Page();
  searchKey: string = undefined;
  sortOrder: string = undefined;
  sortBy: string = undefined;
}